requirements.txt
astroid==1.6.5
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-Cors==3.0.6
freezegun==0.3.10
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
PyMySQL==0.8.1
python-dateutil==2.7.3
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.2.7
sqlalchemy-seed==0.1.2
SQLAlchemy-Utils==0.33.3
Werkzeug==0.14.1
wrapt==1.10.11
